refactor(tax-free-tags): remove unused components and update import paths
